backend/app/core/environmental.py
import httpx
from typing import Dict, Any
import os
from dotenv import load_dotenv

load_dotenv()

async def get_environmental_data(lat: float, lng: float) -> Dict[str, Any]:
    """
    Fetch environmental data from various APIs.
    This is a mock implementation that returns sample data.
    """
    try:
        # Mock environmental data
        return {
            "timestamp": "2024-03-18T00:00:00Z",
            "location": {"lat": lat, "lng": lng},
            "metrics": {
                "aqi": 45,
                "pm25": 10.5,
                "pm10": 20.3,
                "o3": 35.2,
                "no2": 15.8,
                "so2": 5.2,
                "co": 0.8,
                "temperature": 22.5,
                "humidity": 65,
                "precipitation": 2.5,
                "windSpeed": 3.2,
                "windDirection": 180,
                "pressure": 1013.2,
                "ndvi": 0.65,
                "lst": 298.15,
                "waterBodies": 0.15,
                "urbanDensity": 0.45
            }
        }
    except Exception as e:
        raise Exception(f"Error fetching environmental data: {str(e)}") 