backend/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-dotenv==1.0.1
sqlalchemy==2.0.39
psycopg2-binary==2.9.10
pydantic==2.10.6
starlette==0.46.1
typing-extensions==4.12.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
requests==2.31.0
aiohttp==3.9.3
pandas==2.2.1
numpy==1.26.4 